chore(cloudflare): refactor oauth, improve error handling
