+++ alchemy/src/cloudflare/user.ts
@@ -1,7 +1,7 @@
 import type { Secret } from "../secret.ts";
 import { logger } from "../util/logger.ts";
 import { memoize } from "../util/memoize.ts";
-import { CloudflareApiError } from "./api-error.ts";
+import { extractCloudflareResult } from "./api-response.ts";
 import {
   getCloudflareAuthHeaders,
   type CloudflareAuthOptions,
@@ -56,66 +56,48 @@ export interface CloudflareOrganization {
   roles: string[];
 }
 
-export const getCloudflareAccounts = memoize(
-  async (options: CloudflareAuthOptions): Promise<CloudflareAccount[]> => {
+export const getCloudflareAccountId = memoize(
+  async (options: CloudflareAuthOptions): Promise<string> => {
     const headers = await getCloudflareAuthHeaders(options);
-    const accounts = await fetch(
-      "https://api.cloudflare.com/client/v4/accounts",
-      {
+    const accounts = await extractCloudflareResult<CloudflareAccount[]>(
+      "get accounts for authorized user",
+      fetch("https://api.cloudflare.com/client/v4/accounts", {
         headers,
-      },
+      }),
     );
-    if (accounts.ok) {
-      return ((await accounts.json()) as any).result;
-    } else {
-      throw new CloudflareApiError(
-        `Failed to get accounts for authorized user, please make sure you're authenticated (see: https://alchemy.run/guides/cloudflare/) or explicitly set the Cloudflare Account ID (see: https://alchemy.run/guides/cloudflare/#account-id)`,
-        accounts,
+    if (!accounts[0]) {
+      throw new Error(
+        "No accounts found for authorized user. Please make sure you're authenticated (see: https://alchemy.run/guides/cloudflare/) or explicitly set the Cloudflare Account ID (see: https://alchemy.run/guides/cloudflare/#account-id)",
       );
     }
+    if (accounts.length > 1) {
+      logger.warnOnce(
+        [
+          "Multiple Cloudflare accounts found for authorized user:",
+          accounts.map((a, i) => `${i + 1}: ${a.name} (${a.id})`).join("\n"),
+          "The first account will be used by default. To use a different account, explicitly set the Cloudflare Account ID (see: https://alchemy.run/guides/cloudflare/#account-id)",
+        ].join("\n"),
+      );
+    }
+    return accounts[0].id;
   },
 );
 
 export const getUserEmailFromApiKey = memoize(
   async (apiKey: string): Promise<string> => {
-    try {
-      // Call the /user endpoint to get user information
-      const response = await fetch(
-        "https://api.cloudflare.com/client/v4/user",
-        {
-          headers: {
-            "Content-Type": "application/json",
-            "X-Auth-Key": apiKey,
-          },
+    const result = await extractCloudflareResult<{
+      id: string;
+      name: string;
+      email: string;
+    }>(
+      "automatically discover email for API Key authentication",
+      fetch("https://api.cloudflare.com/client/v4/user", {
+        headers: {
+          "Content-Type": "application/json",
+          "X-Auth-Key": apiKey,
         },
-      );
-
-      if (!response.ok) {
-        throw new Error(
-          `Failed to get user information: ${response.status} ${response.statusText}`,
-        );
-      }
-
-      const data = (await response.json()) as {
-        success: boolean;
-        result: {
-          id: string;
-          email: string;
-          name: string;
-          [key: string]: any;
-        };
-      };
-
-      if (!data.success || !data.result || !data.result.email) {
-        throw new Error("Cloudflare API did not return valid user information");
-      }
-
-      return data.result.email;
-    } catch (error) {
-      logger.error("Error retrieving email from Cloudflare API:", error);
-      throw new Error(
-        "Failed to automatically discover email for API Key authentication",
-      );
-    }
+      }),
+    );
+    return result.email;
   },
 );
