+++ alchemy/src/util/memoize.ts
@@ -24,23 +24,6 @@ export function memoize<F extends (...args: any[]) => Promise<any>>(
   };
 }
 
-export function memoizeSync<F extends (...args: any[]) => any>(
-  fn: F,
-  keyFn: (...args: Parameters<F>) => string = defaultKeyFn,
-) {
-  const cache = new Map<string, ReturnType<F>>();
-  return (...args: Parameters<F>): ReturnType<F> => {
-    const key = keyFn(...args);
-    const cached = cache.get(key);
-    if (cached) {
-      return cached;
-    }
-    const result = fn(...args);
-    cache.set(key, result);
-    return result;
-  };
-}
-
 export const defaultKeyFn = (...args: any[]) => {
   return createHash("sha256").update(JSON.stringify(args)).digest("hex");
 };
